Update native package path&remove mapping provider
