+++ ballerina/jsonpath.bal
@@ -17,11 +17,11 @@
 import ballerina/jballerina.java;
 import ballerina/lang.'object as obj;
 
-public type JsonPathValue ()|boolean|int|float|decimal|string|json[]|map<json>;
+public type JsonPathValue json;
 
 public type JsonPathRawTemplate object {
     *obj:RawTemplate;
-    public (string[] & readonly) strings;
+    public string[] & readonly strings;
     public JsonPathValue[] insertions;
 };
 
@@ -34,9 +34,7 @@ public isolated function read(json 'json, JsonPathRawTemplate query) returns jso
 }
 
 public class JsonPathRawTemplateImpl {
-    *object:RawTemplate;
-    public string[] & readonly strings;
-    public JsonPathValue[] insertions;
+    *JsonPathRawTemplate;
 
     public isolated function init(JsonPathRawTemplate jsonPathRawTemplate) {
         self.strings = jsonPathRawTemplate.strings;
@@ -49,5 +47,5 @@ public class JsonPathRawTemplateImpl {
 # + query - JSON path expression
 # + return - extracted details as JSON value, a jsonpath:Error otherwise
 public isolated function readJson(json 'json, JsonPathRawTemplateImpl query) returns json|Error = @java:Method {
-    'class: "io.ballerina.xlibb.jsonpath.BJsonPath"
+    'class: "io.ballerina.stdlib.jsonpath.BJsonPath"
 } external;
