+++ build-config/resources/Ballerina.toml
@@ -11,7 +11,7 @@ export = ["data.jsondata"]
 
 [[platform.java11.dependency]]
 path = "../native/build/libs/jsonpath-native-@project.version@.jar"
-groupId = "io.ballerina.xlibb"
+groupId = "io.ballerina.stdlib"
 artifactId = "jsonpath-native"
 version = "@toml.version@"
