+++ native/src/main/java/io/ballerina/stdlib/jsonpath/BJsonPath.java
@@ -16,7 +16,7 @@
  *  under the License.
  */
 
-package io.ballerina.xlibb.jsonpath;
+package io.ballerina.stdlib.jsonpath;
 
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.JsonPathException;
@@ -27,10 +27,12 @@ import io.ballerina.runtime.api.values.BObject;
 
 import static com.jayway.jsonpath.JsonPath.using;
 
-import static io.ballerina.xlibb.jsonpath.Utils.convertRawTemplateToString;
+import static io.ballerina.stdlib.jsonpath.Utils.convertRawTemplateToString;
 
 /**
  * Provides function implementations of json-path.
+ *
+ * @since 0.1.0
  */
 public class BJsonPath {
     public static final Configuration BJSON_CONFIGURATION = Configuration
